+++ src/server/CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Fixed
  - Fix issue with sending commands
 
+ ### Updated
+ - Add state tracking
+
 ## 1.0.5 (2023-10-11)
 
 ### Updated